Clean code up.

+++ org.jhotdraw8.collection/src/main/java/org.jhotdraw8.collection/org/jhotdraw8/collection/primitive/ByteArraySpliterator.java
@@ -75,7 +75,7 @@ public class ByteArraySpliterator implements SpliteratorOfByte {
 
     @Override
     public long estimateSize() {
-        return (long) (fence - index);
+        return fence - index;
     }
 
     @Override
